Better state handling, notifications
Decided that instead of binding a component against an atom, I'd just
send a notification on change. That way I can keep all state in one place
and clear it if necessary, something that becomes ugly when the state for
the textarea is kept on an atom.

Mike Thompson suggested doing it on-blur instead of updating every
on-change (because of https://github.com/Day8/re-frame/issues/39). While
dispatching on-blur is cleaner, I can't update the textarea to change or
clean it unless I use the `value` attribute... which is blank until we
exit it.  Looks like I have to use `on-change`, so the event needs to be
fired with dispatch-sync.