Setting a thought as its own root when first refined
We are using the root_id as an indicator that a thought is part of a
thread. This lets me show the indicator not only on the follow-ups, but
also on the original, and easily query them all together.

Added migrations to update the data already on the live site.