Performance: converting markdown to html on retrieval
It seems like md->html was a bit expensive when running it on a few
hundred items. It wasn't a dealbreaker, but it bugged me enough to
profile and fix it.

This wasn't an issue before endless scrolling since we always rendered
a limited set of items the moment we queried them.