Avoiding leftover roots on thought delete
When we delete a thought, I'm clearing a thought's root ID if it would
be the last remaining item, to avoid any thread-less roots. Doing it after
delete to avoid potential confusion where we'd need to compare if there
are two remaining items (the one about to be deleted, and the root), or
having an item attempt to update itself.

https://app.asana.com/0/17212341701456/161828115758830