Removing unnecessary clear button on memory
Having it there required a double-tab to get to Submit, and people
could clean it themselves.

Also, the Cancel button doesn't need to be as accessible on Edit, since
we can easily cancel by pressing Escape. Not removing it altogether,
since it could be useful on mobile where there is no escape key.