Added fixtures to mount end and db on every test namespace
I was previously doing this only on a couple namespaces, relying on
clojure.test’s implementation detail that it runs the namespaces in
alphabetical order. That was shoddy, and it broke when integrating
cloverage since it _can’t_ be relied on doing the same.

Removed running migrations on the fixtures, though. They shouldn’t be
introducing side effects. A developer should make sure all migrations
have run before running tests.