BUGFIX on reminder query (#114)
Stupid error on a join, was getting a cartesian product of reminders and
thoughts for a user.

This explains a performance issue I was seeing: I was sending ~1MB of
data in some cases when I could just have been sending 5kb. Doh.

This should also improve performance when querying and reduce memory
consumption.

Expanded tests.